Add make target to embed autonity contract (#481)
This generates the autonity abi and bytecode and embeds them in the
codebase.

The abi and bytecode are no longer committed to git. This helps ensure that we run tests with the correct abi and bytecode, it removes the burden of having to remember to do this step manually after updating the contract.