Allow nil params.User.Address (#515)
This ensures that decoding a genesis file that does not specify a user
address does not result in a struct with a zero valued address.

As a consequence of this change I also needed to update the way we check
for unspecified addresses in the code '== nil' rather than using
reflect.

Also updated the params.User.Validate function to check that if both
user Address and Enode are specified then they both refer to the same
address.

And combined 2 methods of params.AutonityContractGenesis into Prepare()

AddDefault and Validate were being used together in most places and it
seems risky to allow for the possiblity that a caller calls AddDefault
and forgets to call Validate. Now they are merged that cannot happen.
This change highlighted a number of places in our tests where validation
had not been being applied correctly and so warranted a number of test
fixes.