Improve resiliency of test-contracts make target
The target was failing in ci when autonity was taking too long to
start up, this adds a loop to retry connecting to autonity up to 10
times with a 2 second gap between attempts.

This also moves the installation of truffle and web3 into the make
target so that it is more self contained.