Use Node.Close() instead Stop() for force stop
Close() cleans up resources allocated in New and so avoids leaked
goroutines.

As a consequence needed to ignore the TestTendermintStartStopXXXX tests.

The effect of calling Close() instead of stop is that the
TestTendermintStartStopXXXX tests fail because the second call to
Close() results in the accounts.Manager trying to close an already
closed channel. The account manager is created in node.New and so in
order to overcome this problem we will need to fully recreate the node
in our tests. I think that is quite a large change and is probably out
of scope for this PR.