Validate that difficulty is 1 in genesis file (#510)
Previously we were setting the difficulty to 1 during genesis block
generation if we detected that the autonity/tendermint config had been
set.

After a discussion, see here:
https://github.com/clearmatics/autonity/pull/510#discussion_r417539836

We decided to instead validate that the difficulty is 1, so that users
who may have set a difficulty other than 1 in their genesis file would
be alerted to their error rather than have the value silently changed.