Separate unexported comittee methods from backend.Backend
This is a first step in a strategy to be able to encapsulate the
proposer selection logic inside the committee set. Currently we have a
single committee set implementation that provides a plain round robin
proposer selection. The weighted proposer selection however has it's
logic spread out between tendermint/core.core tendermint/core.BlockChain
tendermint/backend.Backend and tendermint/committee.Set.

The top level where the decision between using a weighted proposer
selection vs simple round robin is in tendermint/core.core. So we want
to be able to isolate comittee.Set creation to there since core.core has
all the required knowledge to make this decision and should hence have
(or be able to collect) all the required parameters.

I can see that tendermint/core.core's committee.Set is being constructed
in setInitialState and is being provided by a tendermint/core.Backend
instance which is actually an instance of tendermint/backend.Backend.

So as a starting step I will try to separate the committee.Set creation
logic from tendermint/backend.Backend. It turns out that the committee
and savedCommittee methods on tendermint/backend.Backend don't actually
make use of any other fieds or methods of Backend and so can be moved
into standalone functions.