Removed backend.Backend.Committee() and core.Backend.Committee()
Replaced calls to the Comittee methods removed with calls to
SavedCommittee. I needed to move SavedCommittee to its own package to
break a circular dependency between tendermint/core and
tendermint/backend arising from use of backend in
tendermint/core.core.setInitialState. I chose package temp because this
is not the final design SavedCommittee is just there till the
simplification can be completed, then we can find a proper home for it.

This change helps to untangle things from the backend package and the
backend.Backend struct, since we can now get a saved committee without
depending on backend or backend.Backend.

As part of this I needed to expose a method to get a BlockChain from the
Backend.