Fix tendermint/backend.Backend.committee function
The committee for each block is defined in the parent header, the
committee function was using the committee defined in the current header
as the committee for the current block. This change ensures that the
committee function correctly uses the committee defined in the parent
header as the committee for the current block.