Pass required parent instead of slice of parents
A number of functions in tendermint/backend were accepting a slice of
*types.Header holding parent headers but only using the last header,
this is confusing. This change instead only passes the parent header to
functions that require it as opposed to a slice of parents.

This change should help in determining if we can replace
temp.GetCommittee with a simpler function in certain places.