Combine the 2 code paths in GetCommittee into one
The previous implementation of GetCommittee behaved differently depending
of whether a parent header was provided. It now behaves in the same way
apart from looking up the parent if missing. Also added a quick exit
code path in the case that the provided header is the genesis header.

This allowed me to simplify the code in verifyCascadingFields to just
look up the parent and pass it on.