Remove use of GetCommittee in verifyCommittedSeals
verifyCommittedSeals was calling temp.GetCommittee but not using the
full functionality of committee.Set.

Since verifyCommittedSeals now has a mandatory parent header parameter
we can rely on that to get the committee members and then we just needed
a way to calculate the quorum threshold, which I separated into its own
function in tendermint/committee

Not calling GetCommittee removed the requirement for having a chain
parameter and allowed us to remove the chain parameter from a hierarchy
of 3 functions.