Move parent header retrieval up call hierarchy
Moved parent header retrieval into VerifySeal which is the top level of
the consensus.Engine interface.

This allowed me to simplify verifySigner, it no longer needed to call
temp.GetCommittee because the information it needed was available on the
parent header.

Removing use of temp.GetCommittee allowed the removal of the chain
parameter in verifySigner.