Remove use of temp.GetCommitteeSet in backend_test
For some reason we were creating a committee.Set to set the private key
of the proposer for the backend, but the private key was already being
passed to the backend and in the case of the genesis block we always
know the proposer will be the first member of the committee.

Also removed methods to set and get private key on the backend.