Move committee into core package
Provide 2 implementations of comittee.

This allows for the committee implementation to be chosen by core.core based on
the config provided to core.core.

As well as moving the creation of the committee object into the core package I
also decided to move the implementations into the core package, since the only
place we were using the logic to determine the proposer was in the core
package.

Some tests in the backend were creating committee instances used as parameters
to Backend.Broadcast and Backend.AskSync I was able to replace these with
instanes of types.Committee and types.Header. AskSync was also making use of
the Quorum method from the committee so I separated that out into a standalone
function in order to remove the dependncy on a committee instance. It seems to
me that calculation of quorum size and tolerable failures is sufficiently
different from proposer selection that it is reasonable to represent them as
separate objects.

Also crypto.CheckValidatorSignature was accepting a committee instance which
was easily replacable with a types.Header instance.

By limiting the use of committee to the core package I was able
shrink the committee.Set interface.