Created Autonity Contract and FullChainReader interfaces
In order to untangle the import dependencies in the core package I introduced the FullChainReader interface
and changed the signatures of the state full-access needed by some of the consensus functions.
I introduced as well an interface for the Autonity Contract to make testing possible from the miner package.
This would make easier I believe to perform an eventual switch to Autonity without the evm contracts.

Signed-off-by: yazzaoui <ya@clearmatics.com>