Run init step in now failing tests
Some of the autonity cmd tests were not initing the client before
running it, this commit ensures that they are initing themselves
correctly. Unfortunately the tests are now failing because they are
making requests to the tendermint api causing the backend.Backend
instance to try and acccess a nil core.BlockChain field.