Fixed crash but now tests don't complete
This work is now blocked.

Initing autonity requires valid autonity genesis, see:

cmd/autonity/chaincmd.go:220
utils.Fatalf("No Autonity Contract config section in genesis")

Initing autonity causes the console to try to request the autonity
contract, but it has not been deployed.

Waiting for the first block to be mined does not work, it is never
mined. I'm not sure why but getting it to be mined in the test
environment looks hard, since the client is being started as an external
process.

I think the best approach now is to deploy autonity on init. And rewrite
these tests in a more sane way. (its totally unnecessary to start an
external process and scrape its standard out in order to check that the
console dipslays the right data)