Hardcode autonity contract and deployer address (#552)
No user is allowed to make calls to autonity contract thus the need for a deployer address which is used to ensure only the client can make restricted calls to autonity contract. 

However, this deployer address can be set during the genesis and if anyone owns the private key to this address, they can easily change the state of the contract. Thus this commit ensures that there is no option to set the deployer address instead it is hardcoded to "0x0" (known as the black hole address in the ethereum community as it is widely believed no one has the private key for this public address). By hardcoding the deployer address, the autonity contract address can also be hardcoded since it is calculated deterministically from the deployer address.