Hide the config bytecode and abi fields from users
These fields are part of the Autonity contract config but in order to
correctly set them users would have to have, a very in depth
understanding of how Autonity is working. Until we can provide the
support required for users to know enough to set these fields correctly,
we hide them, by removing documentation for them and ensuring they do
not appear in serialised genesis files.

Closes #513