Remove config for proposer selection mechanisms
This is the first step towards removing the round robin implementation.

Because of this core.TestHandleCommit is failing because we are
constructing the committee set inside core and the weighted random
sampling set needs to use a blockchain to get a statedb, to make this
work through mocks we would have to mock a blochchain to return a
statecache that could be used in a real evm. This would be a lot of
work.

The much simpler approach to this is to just construct core with a
function that when given the height and round returns the proposer.

EG: func(height, round int64) types.CommitteeMember

Mocking this will be so simple we won't even need to use a mocking framework.

This is the approach I will take.