Wip fixing errors in upon_condition_test.go
The upon condition test is extremely fragile, since it is constructing
large heirarchys of objects with many parts missing or mocked but not in
a consistent or cohesive way, so the changes made to core in order to
add weighted random sampling are going to require a lot of changes to
these tests which are not obvious and will require a lot of trial and
error.

I have made some progress fixing these errors but Its too much to
finish before I leave. The best solution here would be to untangle the
code under test and then clean up the tests, if we do not do that I can
imagine these tests taking a very long time to fix whenever core is
changed.