Prevent unintended test runs in ci
When using a github actions matrix to run tests we were specifying
specific tests to run by name. Using this approach 'go test' will run
all tests with a matching prefix, which  in our case was causing us to
run tests more than once.

For example the following 2 matrix entries would cause
TestTendermintMajorityExternalUsersFullyConnected to be executed on each
run.

  TestTendermintMajorityExternalUsers,
  TestTendermintMajorityExternalUsersFullyConnected,

This commit updates the matching regex to match the word boundaries at
the beginning and end of the test name so that we only run the specified
test.