Add detailed panic to provide better feedback #624
The panic mentioned in
https://github.com/clearmatics/autonity/issues/624 is occurring when the
second parameter to big.Int.Add is nil which is in turn caused by
blockchain.GetTd returning nil for the parent of the current block. The
panic happens rarely and we need more knowledge of what is the current
state to better understand how it is ocurring.

This commit adds a much more detailed panic that includes a dump of the
child block and the specific arguments in the call to blockchain.GetTd.