Remove newlines in generated code echo statements #622
Bash's echo does not inerpret '\n' as a special character unless '-e'
flag is provided and sh's echo does not understand the '-e' flag.

This meant running 'make embed-autonity-contract' when the default shell
is bash would ouptut the '\n' into the generated code.

This commit removes the use of '\n' in the echo  statements used to
generate code. We now use 'gofmt -s' to ensure the files are properly
formatted, this required updating the bootstrap build job in our github
actions so that golang is available.