Remove use of docker during 'make autonity' #622
Instead of using the dockerised solc to compile the contract we use the
solc binary. This allows us to run 'make autonity' during docker build,
which in turn means that we can build autonity entirely within docker.
The reason for making this change is that devops requested that they be
able to build the docker image on a host with only docker installed.
Previously we would have required a host with make in order to run
embed-autontiy-contract before running the docker build.