fix freezer hash comparison by correctly calculating header hash
freezer stores raw rlp values and go-ethereum's header hash is
calculated by keccak256(rlp(header)), which is not true for autonity
header hash, where few fields are remove before creating the rlp and
the hash (please see core/types/block.go:155). Therefore, when freezer
reads raw rlp values from disk it needs to ensure that the hash matches
the one being requested (since forks can happen in go-ethereum and for
the same block number there can be different hashes, which is shouldn't
be the case in autonity, thus this can be seen as redundant) we need to
calculate the autonity header hash before comparison. This can only be
done be decoding the header and then creating its hash. The caller of
ReadHeaderRLP will most likely decode the header again which is not
efficient.