replace *prque data structure with []*Message to manage backlog
*prque has a memory leak where the heap size is never shrunken even
after the elements are popped from the priority queue. This commit
uses slice to manage the backlog where for each committee member a slice
of message is created and messages are added to the slice when
storeBacklog() is called. Slices can not keep the concept of a priority
therefore during processBacklog() the whole slice needs to be iterated
and checkMessage() is called to determine whether it needs to be sent to
the handler. This is the only difference between the priority queue and
slice implementation in this context. The priority queue can stop the
processing as soon as a future message is encountered where as with the
slice the whole needs to be iterated since it is not sorted.

We also need to ensure that the underlying array which the slice is
pointing to never becomes very large in capacity while the length of
the slice is small. This can lead to a memory leak, therefore, this
commit explicitly allocated new slice when capToLenRatio is exceeded.