Make nodes wait for transactions to mine a block
Also upped the verbosity for the test node so that I could visually see
the logs for block creation.

Unfortunately this does not work with the test framework since it
requires a block to be mined before the nodes start to send
transactions.

Also tweaked the blocksToWait variable. In anticipation of a cleaner
testing approach.