Removing now uneeded code
Realised that the existing implementation is predicated on the idea that
we process prevote and precommit messages only after we have moved on
from the propose step and that when we process the backlog we only
process the next step's worth of messages. This approach breaks the
algorithm or at least means that we are not correctly following the
algorithm.

The better approach would be to process all messages when they arrive
and each message faces all the checks for upon conditions.