Deploy autonity contract on init (#611)
This commit separates the contract deployment method from the autonity
struct, allowing it to be called without needing to satisfy all the dependencies
of the autonity contract. It also introduces a method to construct an evm at the
genesis state. With these 2 methods we were able to deploy the autonity
contract on init and remove the deployment during processing of block one.
Finally it fixes all the tests that broke as a consequence of this change.