fix freezer hash comparison by correctly calculating header hash (#639)
freezer stores raw rlp values and go-ethereum's header hash is
calculated by keccak256(rlp(header)), which is not true for autonity
header hash, where few fields are removed before creating the rlp and
the hash (please see core/types/block.go:155). Therefore, when the freezer
reads raw rlp values from disk it needs to ensure that the hash matches
the one being requested (since forks can happen in go-ethereum and for
the same block number there can be different hashes, which it shouldn't
be the case in autonity, thus this can be seen as redundant) we need to
calculate the autonity header hash before comparison. This can only be
done by decoding the header and then creating its hash.

ReadHeaderRLP returns both the encoded and decoded header values
This is in order to prevent decoding the value twice, therefore, it is
up to the caller to determine how to use these returned values