Unlock accounts at autonity start
This commit ensures that we start autonity with flags to unlock required
accounts as opposed to requesting account unlocks via the autonity
console.

It looks like when requesting an account unlock from the console, the
console returns while unlocking continues in the background, so you
can't know when the account will be ready.

It also looks like a similar thing occurrs when requesting unlocks via
the commandline, because the ethereum node is started before accounts
are unlocked (see link below), so this is by no means guaranteed to
solve the issue. But it seems to have stopped the failure seen in #626
from ocurring and so is an improvement.

https://github.com/clearmatics/autonity/blob/a5d4c6bd8d607a413c41220f0812cdd3f07adece/cmd/autonity/main.go#L323-L327