Add initial cobra cli
Regenerated the genesis in the helloworld example project so that, the
operator account uses nodekey1 as the private key, which helped with the
testing.

Also commented out code to deserialise commited seals in
gen_header_json. There is a problem with the deserialisation code, but
we don't depend on the committed seals in this tool so its ok for now.