Re-worked gengen add keys to the user string
Previously it accepted a separate file with keys, but this was a bit
clunky becuse you needed to either start with no keys or have a key per
user in the file.

Now you can generate a genesis with some users having a key and others
having theirs generated at construction time.

Also changed the key format, so that there is no 'priv:' or 'pub:'
prefix and insted the type of key to load from the file is determined by
the length of the hex string. This means keys should work with the
existing tooling which just expect the hex encoded pub or private key.