Fix marshalling
We had a mismatch between the json rpc marshalling and the basic json
marshalling of types.Header, now both situations use the marshalling
defined on types.Header.

This also fixes the marshalling on types.Header. By adding marshalling
to hex for CommitteeMembers and Round.

Additionally the json marshalling was confused because it was packing
and unpacking fields into extra data when that was not required.

The approach of unpacking and packing into extra data is required when
it comes to EncodeRLP/DecodeRLP because rlp encoding is not able to
ignore fields it does not recognise, and so if we wish to maintain rlp
compatibility with existing ethereum tooling we need to make sure that a
serialized header contains exactly the same type and ordering of fields
as the official ethereum header.

Also updated the rlp unmarshaling so that we do not duplicate the extra
header data by unpacking it into fields and also leaving the raw extra
data set on the header.