Add node console (#676)
* Add node console

Added a script that starts up the nodejs console with a web3 instance
connected via the provided websocket rpc address. It also provides
bindings for the Autonity contract.

* Add web3 extensions based on loaded modules

We introspect the loaded rpc modules at startup and apply the
appropriate web3 extension based on the loaded modules.

Now that we have the tendermint web3 extensions we are able to
dynamically get the autonity contract bytecode from the service for the
bindings, instead of relying on the generated file inside the repo. This
makes the 'nodeconsole/console' script very portable.

The code that we were porting over to handle the web3 extensions was
very old and the web3 library has moved on from there, therefore it has
no concept of extended properties, we will need to see how we handle
those, for now they are commented out.

* Remove step to genrate autonity contract abi

We now request it from the node.

* Move console to nodejsconsole

This helps to distinguish it from the generic meaning of node, which we
use to mean running aotonity client.