Cleaned up console init code
Now we start the repl explicitly which means we can do that after our
startup code, ensuring that the prompt appears properly at the start of
the console.

Also removed use of the closure, by starting the repl explicitly we get
to set what appears in it's global scope.

Also cleaned up indentation