Solved ws issue & a little progress on start stop
The websocket subscription problem was a non issue. When a node awaits,
it waits for all sent transactions, I was not clearing out the sent
transactions when they were processed, meaning that subsequent awaits
would wait forever.

Made a little progress on the start stop tests but still stuck because
of this issue, how to manage the outstanding transactions nicely. Maybe
go back to my initial approach of opening a subcription to start with
and storing all processed transactions in a map. That way I can track
transactions in the node and just keep adding to the list of sent
transctions, and then when I call await transactions it checks them all.
Preverified transactions will be very quick to verify becase they will
be in the map.

This covers the case I'm currently struggling with where a call to await
times out but we later want to check that the tx has been processed.