Fix non progression issue
I was not signalling unles c.value was nil, but I was forgetting to
clear c.value when awaiting a new value and the current value was not at
the correct height. This meant that the waiting routine was never
signalled, because the notifying routine only notified if the value was
nil.