Remove sanity checks of new unmined blocks
The test is basically working except it doesn't close, need to work the
context into the awaitValue function.

The awaitValue method checks that it is getting a value for the right
height, so we don't need to check at the new unmined block point.

Because the unmined block processing and the setting of height occur in
different goroutines you can't make any assumptions about the timing of
events coming through, so the case I was panicking on is actually totally possible.

Could this be made simpler by just integrating this into the main loop.
Maybe not, because if we did that what would we do when an event kicks
off that is meant to start the next height but we don't have avalue. We
would just not start the next height, and then when a value came we
would start.