Change algo to not wait for value
We now pass value directly into the algorithm and we wait for it
outside, this had the result of needing to always call StartRound from
outside of the algorithm so that a value can be passed in from outside.

The benefit of this is that algorithm becomes even simpler. All its
methods are now non blocking.