Separate out state dump tests
The tests as they were shared a huge context (in the sense of variables
and closures in scope) and that made it very hard to see what each test
was actually using out of the common context. This commit moves each
test to a separate test which dramatically shrunk the context for most
of the tests making them much easier to follow.