Update linting diff commit
Now we use the merge base between develop and the branch being linted as
opposed to just diffing the branch with develop. This prevents linting
including changes introduced on the develop after the point that your
branch started.

In order to support this the checkout for lint on the CI is no longer
shallow.

Fixes #696