Ensure miner worker test generates valid genesis
The changes in this branch caused us to try to get the proposer where
before we were not. The genesis block that was previously being used for
these tests was not a valid block for tendermint and had no committee
and so resulted in an error when trying to find the proposer.