Broadcast decided block instead of passing it to the fetcher
The fetcher Enqueue method basically does a whole load of validation on
the block to ensure that it is good before broadcasting it, but because
we just decided on this block we don't need to go through all that
validation again. The path the block takes through the fetcher is also
horribly complex.