Add a script for running the Docker E2E tests locally in a VM
The Docker end-to-end tests for Autonity were primarily designed to be
executed in a CI system like GitHub Actions, however it can sometimes be
beneficial to run them locally for retesting or analysis.

As the tests are executed with superuser privileges, they could be
dangerous to run on a workstation. Also, because are assumed to run in
Ubuntu, they could be fiddly to set up on any other system.

This commit adds a Vagrantfile for setting up an Ubuntu virtual machine
and a shell script for running the tests inside the VM.