Fix committed block handling
The stop chan parameter to the Seal function is a per sealing operation
parameter, not an indication of whether the miner has been stopped. So
we now open a new goroutine for each seal operation to read from the
commitChannel and write to the resultsChannel.

Also the commitChannel was not being made, this is now fixed.