Ensure that tendermint algo state not lost
When stopping and restarting we were overwriting the tendermint
algorithm round and step. This allowed restarted nodes to lock a value
and then forget it, which allowed more than 1/3 of the network to be
locked on a value with less than 2/3 of the network locked on that
value, which leads to a state where no new proposal can ever be agreed
upon.