Simplify github actions go workflow
Not including comments this cuts the size of the workflow file from 425
lines to 196.

It also introduces the go cache which holds both module
source and build artifacts, this significantly speeds up building in
subsequent runs.

We also do not refer to tests by name any more, which means we do not
need to worry about forgetting to add new e2e tests to the workflow
file. We do however split tests by package so that we can shorten the
overall running time of the workflow, so when those package names change
we will need to update the workflow.

To support the simplification the makefile has been updated to make sure
that targets rely on their dependencies. E.G: The contract tests now
will ensure that autonity is built before trying to run an aoutonity
instance.