Ensure that core doesn't deadlock when closing
When closing we could be in a situation where the go routine that
pulled from the unmined block channel would exit, and the go routine
that pushed to the channel would get stuck. Since core is waiting for
all routines to exit before it closes this caused a deadlock when
closing.

This is fixed by ensuring that the go routine that pushes to the unmined
block channel can break from that when core is closed.