Fix header marshalling
Header marshaling was broken because there were 2 different and
conflicting ways of marshaling/unmarshaling to JSON. One in

autonity/internal/ethapi/api.go RPCMarshalHeader

And the other in

core/types/gen_header_json.go

Additionally we had manually modified core/types/gen_header_json.go to
encode our additions to the header structure to and from extraData when
marshaling to and from JSON. This was un-necessary since adding extra
fields to JSON object does not break JSON marshaling and it also meant
extra work for clients using our JSON APIs. It was another layer of
decoding for them, and how could they be expected to decode it given
that we provided no specification about the format of what was encoded
in the extraData. It was also painful for us to maintain since we could
no longer use gogenerate to generate the JSON marshaling code.

Also we were not encoding CommitteeMembers as hex which seems to be a
standard for ethereum(no idea why!).

This change adds hex encoding for CommitteeMembers.

It removes any custom adjustments to core/types/gen_header_json.go and
regenerates it with gogenerate, this results in the JSON header
representation directly containing our extra header fields.

It replaces duplicated marshalling in autonity/internal/ethapi/api.go
with the marshaling from core/types/gen_header_json.go.

When marshaling to/from RLP we still encode our extra fields inside the
extraData field since RLP is sensitive to the number of fields in an
object and trying to decode/encode an object with more fields than
expected will fail, so this should ensure compatibility with tools that
want to work with RLP encoded data.