Add new test package
The test package makes it easy to set up, interact with and tear down
networks of nodes.

It is intended as a replacement for our existing test framework, which
because it was a framework required you to fit into its structure. This
restricted what could be done and it also obfuscated what was happening
in the tests.

This commit also adds some tests to demonstrate use of the new test
package, some of these tests replace tests from the old framework.