Fix slow syncing
Sometimes tests would be very slow, this was due to the sync not
happening correctly. When nodes would start they would request to sync
sometimes before any peer had connected, and this would result in them
not syncing at that point and waiting for the next sync interval (10s).
Now if nodes do not see any peers when attempting to sync they will
sleep for 10ms and try again, until they have peers to sync with.

Additionally nodes were slow to perform a blockchain sync, this was due
to the 'eth.defaultMinSyncPeers' being set to 5, meaning that blockchain
sync would not occur in response to peer changed events unless we had at
least 5 peer connections. This meant that blockchain sync was happening
only on the force sync timer which was again 10 seconds. Since most of
our tests have 5 or less nodes we would almost always face this slowdown
in our tests. Now 'eth.defaultMinSyncPeers' is set to 1 which means that
nodes will try to sync the blockchain as soon as they have a single
peer.