Remove committee interface and implementations
The committee was over complicating the code of the bridge. The
committee interface was also unnecessary since all but one of its
methods were only used in tests.

As part of this change I was able to get rid of the hack to avoid
calling the contract at genesis, this also required updating the
Autonity contract constructor to ensure that the committee is created at
deployment.