Make tendermint process messages from committee
This fixes a bug where messages stored for a future height would factor
into the vote calculations when that height became the current height.

This was because when we re-rprocesed those messages we handle each
message in turn and remove it from the store if it was not part of the
committee. But if a message was from the committee it would be processed
in the context of other messages (not yet processed) that may not be
from the committee.