Fix block propagation
We were exiting from trying to pass the committed block back to the
miner if the miner closed the stop channel. But this would leave us
stuck, since we would have mined a block, and therefore would be asking
the block awaiter for the next height block, but the miner would not
know about the mined block and instead keep passing us blocks at the old
height to seal, these would never get picked up by the block awaiter.

The fix is to ensure that once a block has been confirmed we either pass
it to the miner or give up when the bridge is shutdown.