Reinstate TestTendermintBusSuccessWithExternalUser
I believe it was failing because
A) We had stopped rebroadcasting and that will break bus tests and
B) External users don't mine after the miner stops them to sync, because
they are not able to sync, because they are not in the whitelist. The
test framework was pausing waiting for all nodes to start mining, but
that was destined to fail.