Simplify core
This commit completely rewrites the core of the system, it removes both
the core and backend packages and replaces them with a single object
Bridge. The bridge bridges the gap between the ethereum system and the
now new implementation of the tendermint algorithm in the 'algorithm'
package.

A lot needed to change to accommodate this but the result is
significantly less code (minus 2250 lines, not including tests) and a
much less tangled system that is easier to reason about.

Also the new tendermint algorithm implementation has very few
dependencies making it much easier to test.