nodejsconsole: drop the use of package-lock.json
The `package-lock.json` is not actually used to fix the dependency
versions. Every time the script is run for the first time, the latest
`web3` is being fetched and used.

Therefore, this file does not add a lot of value. We don't need to
create it, and the existing one can be deleted. Dependencies can be
managed in the `console` shell script directly.

Also, because this `console` script manages dependency installation
automatically, by symmetric logic it should also manage dependency
updates automatically to maintain `web3` at the latest version.

If we need to fix the web3 version in future, it can be done in the
script itself.

The terminal output was streamlined to reduce noise, as npm will tell
you what it is doing if it needs to do something.