Revise the issue templates for Autonity
We can customise the template chooser using a `config.yml` file. In this
way we can add additional links to the New Issue page to help direct
people to the right place.

This commit adds links to the Discord for support, to the AEPs site, and
the docs repo for doc issues.

We also drop the two emoji because they don't add very much and it is
not possible to style the security vulnerability link, which is
automatically created by Github.