Test the ASM contacts using Ape Framework
This change adds a new Makefile target to test the ASM contracts.

The tests make use of the Ape Framework, which supports a number of
different EVM implementations including Hardhat and Foundry's Anvil.

Although it ships with a default EVM provider, this provider does not
support tracing, so it is necessary to install another provider for
testing the ASM.

The current Makefile target uses Hardhat because it is easier to install
in a reproducible way with minimal disruption to developers.