Revise ASM SupplyControl Contract construction
The `operator` of the SupplyControl is the address that is authorized to
mint and burn Auton supply, and is intended to be the Stabilization
Contract.

With this change, the SupplyControl contract takes an additional
argument for the `operator`.

Previously, the SupplyControl Contract was not passed the address of the
`operator` at construction time, because it was assumed that the address
of the latter would not be known at that time in the deployment
sequence. So it was necessary to set the operator of the SupplyControl
in a subsequent call to the contract to avoid circular dependency.

However, as the ASM is deployed at genesis, and the addresses of all the
protocol contracts is known at that time, there is no loneger any need
for the two-step process.