Pass operator and oracle to ASM contract constructors
In a recent commit, the ASM was modified so that the ACU and
Stabilization contracts only receive the address of the Autonity
Contract, and they don't receive the address of the Oracle or the
Operator directly. Instead, if they need to access the oracle or
operator address, they call `autonity.getOracle()` or
`autonity.getOperator()`.

Unfortunately, it was found that this tight coupling between the
Autonity Contract and the dependent ASM contracts makes the ASM more
difficult to test, even with mock implementations of `IAutonity`.

The main issue is the complexity of setting up the Oracle Contract in a
specific state, and then being able to call `finalize()` on it. This
will only become more tricky in future as we will need to do the same
for the ACU as well.

Instead of the indirection, this change passes the Governance Operator
address and the Oracle Contract address directly to the ASM
constructors. The downside is that we have to explicitly update these
contracts when there is a change to the operator or oracle. But at this
time it seems to be a smaller price to pay in terms of development and
maintenance effort.